# internal/web/openapi/openapi.v1.yaml
openapi: 3.0.3
info:
  title: openfsd Operator REST API
  description: |
    JSON control surface for **operator automation** against an openfsd deployment.

    **Blast radius:** Admin-minted API tokens currently carry Administrator-equivalent
    network rating claims. Treat tokens as full operator credentials until fine-grained
    scopes exist. Prefer short TTLs (max 90 days) and rotate via secret reset when needed.

    **Versioning:** URL major is `/api/v1`. Rare breaking changes use the
    `OpenFSD-API-Version` date microversion header (`YYYY-MM-DD`). Production clients
    MUST send a supported pin. When omitted, the server defaults to `max_version` and
    sets `OpenFSD-API-Version-Defaulted: true`. Discovery (`GET /api/v1` and
    `GET /api/v1/versions`) and OpenAPI are version-agnostic (never reject on pin).

    Envelope for JSON resources: `{ "version": "v1", "err": string|null, "data": … }`.
    Body `version` is the major only; microversion appears only in response headers.
  version: "1.0.0"
  x-openfsd-api-micro-min: "2026-07-28"
  x-openfsd-api-micro-max: "2026-07-28"
servers:
  - url: /api/v1
    description: Relative to the openfsd web listener
...
paths:
...
  /editor/validate-apt:
    post:
      tags: [Editor]
      summary: Validate APT text (Instructor1+)
      operationId: postEditorValidateApt
      x-openfsd-stability: stable
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/OpenFSDAPIVersion"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                text: { type: string }
      responses:
        "200":
          description: Soft validation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIV1Envelope"
  /editor/validate-air:
    post:
      tags: [Editor]
      summary: Validate AIR text (Instructor1+)
      operationId: postEditorValidateAir
      x-openfsd-stability: stable
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/OpenFSDAPIVersion"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                text: { type: string }
      responses:
        "200":
          description: Soft validation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIV1Envelope"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access JWT (API token or short-lived access). No CSRF required.
    cookieAuth:
      type: apiKey
      in: cookie
      name: openfsd_session
      description: First-party session cookie; mutations require CSRF synchronizer.
  parameters:
    OpenFSDAPIVersion:
      name: OpenFSD-API-Version
      in: header
      required: false
      description: |
        Date microversion pin (`YYYY-MM-DD` or `1.YYYYMMDD` or `latest`).
        Production clients MUST send a supported pin. Omitted → max + Defaulted header.
      schema:
        type: string
        example: "2026-07-28"
  schemas:
    APIV1Envelope:
      type: object
      required: [version, err, data]
      properties:
        version:
          type: string
          enum: [v1]
          description: URL major only; microversion is header-only
        err:
          type: string
          nullable: true
        data:
          nullable: true
...